Changed how database stores pledges and payments so that there is no more bug of extra digits on user and job detail pages.

+++ jobuser/models.py
@@ -6,20 +6,20 @@ from datetime import datetime;
 class JobUser(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE);
     job = models.ForeignKey(Job, on_delete=models.CASCADE);
-    amount_pledged = models.PositiveIntegerField(default=0);
-    amount_paid = models.PositiveIntegerField(default=0);
+    amount_pledged = models.FloatField(default=0);
+    amount_paid = models.FloatField(default=0);
     oldest_work_date = models.DateTimeField(default=None, null=True, blank=True);
     newest_finish_date = models.DateTimeField(default=None, null=True, blank=True);
-    amount_received = models.PositiveIntegerField(default=0);
+    amount_received = models.FloatField(default=0);
 
 class Pledge(models.Model):
     jobuser = models.ForeignKey(JobUser, on_delete=models.CASCADE);
-    amount = models.PositiveIntegerField(null=True, blank=True);
+    amount = models.FloatField(null=True, blank=True);
     date = models.DateTimeField(auto_now_add=True);
 
 class Pay(models.Model):    
     jobuser = models.ForeignKey(JobUser, on_delete=models.CASCADE);
-    amount = models.PositiveIntegerField(editable=False, null=True, blank=True);
+    amount = models.FloatField(editable=False, null=True, blank=True);
     receiver = models.ForeignKey(User, on_delete=models.CASCADE);
     date = models.DateTimeField(auto_now_add=True);
